requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
wandb>=0.15.0
tqdm>=4.65.0
